requirements.txt
python-dotenv
colorlog

mongoengine
flask
flask-smorest
marshmallow

dash
dash-bootstrap-components
plotly

PyYAML
Werkzeug
WTForms
